requirements.txt
autopep8==1.5
prompt-toolkit==1.0.14
pycodestyle==2.5.0
Pygments==2.5.2
PyInquirer==1.0.3
pytube3==9.5.8
regex==2020.1.8
six==1.14.0
typing-extensions==3.7.4.1
wcwidth==0.1.8
